requirements.txt
ipykernel
python-dotenv
langchain
langchain_community
langchainhub
pypdf
bs4
arxiv
pymupdf
langchain-text-splitters
langchain-openai
chromadb
sentence-transformers
faiss-cpu
streamlit
groq
langchain_groq
langchain_core
fastapi
langserve
uvicorn
langchain_huggingface
arxiv
wikipedia